refactor(habitat): classify workspace model by kind
