+++ tools/habitat-harness/src/service/model/workspace/helpers/diff.helper.ts
