+++ tools/habitat-harness/src/service/model/workspace/policy/target-plan.policy.ts
@@ -1,6 +1,9 @@
 import type { WorkspaceProject } from "@internal/habitat-harness/providers/nx/schema";
-import type { ClassifiedTarget, UnavailableClassifiedTarget } from "./schema.js";
-import { projectTargetStates, workspaceTargetStates } from "./states.js";
+import type {
+  ClassifiedTarget,
+  UnavailableClassifiedTarget,
+} from "../dto/classify.schema.js";
+import { projectTargetStates, workspaceTargetStates } from "./workspace-targets.policy.js";
 
 export function projectTargets(project: WorkspaceProject): {
   targets: ClassifiedTarget[];
