+++ tools/habitat-harness/src/service/modules/classify/contract.ts
@@ -1,5 +1,5 @@
 import { habitatServiceErrorMap } from "@internal/habitat-harness/service/errors";
-import { ClassifyResultSchema } from "@internal/habitat-harness/service/model/workspace/schema";
+import { ClassifyResultSchema } from "@internal/habitat-harness/service/model/workspace/dto/classify.schema";
 import type { HabitatServiceProcedureContract } from "@internal/habitat-harness/service/procedure-contract";
 import { toStandardSchema } from "@internal/habitat-harness/service/typebox-standard-schema";
 import { eoc } from "effect-orpc";
